feat: 예약이 존재하는 시간 삭제 불가 처리

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,6 +3,7 @@ package roomescape.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.domain.ReservationTime;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.service.dto.request.ReservationTimeCreateRequest;
 import roomescape.service.dto.response.ReservationTimeResponse;
@@ -14,6 +15,7 @@ import java.util.List;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
     public List<ReservationTimeResponse> getTimes() {
         return reservationTimeRepository.findAll()
@@ -34,6 +36,11 @@ public class ReservationTimeService {
     }
 
     public void delete(final Long timeId) {
+        final boolean hasAnyOngoingReservation = reservationRepository.existsByTimeId(timeId);
+        if (hasAnyOngoingReservation) {
+            throw new IllegalArgumentException("해당 시간대에 잔여 예약이 존재합니다.");
+        }
+
         final boolean deleted = reservationTimeRepository.delete(timeId);
 
         if (!deleted) {
